LayoutTests/http/tests/xmlhttprequest/access-control-preflight-credential-async.html

<html>
<body>
<p>Test case for bug <a href="https://bugs.webkit.org/show_bug.cgi?id=37781">37781</a>: [XHR] Cross-Origin synchronous request with credential raises NETWORK_ERR</p>
<pre id='console'></pre>
<script type="text/javascript">
function log(message)
{
    document.getElementById('console').appendChild(document.createTextNode(message + "\n"));
}

if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

try {
    var xhr = new XMLHttpRequest;
    xhr.open("PUT", "http://localhost:8000/xmlhttprequest/resources/basic-auth/access-control-auth-basic.php?uid=fooUser", false, "fooUser", "barPass");
    xhr.onerror = function (e) {
        log("FAILED: received error");
        if (window.layoutTestController)
            layoutTestController.notifyDone();
    };
    xhr.onreadystatechange = function () {
        if (xhr.readyState == 4) {
            log((xhr.status == 401) ? "PASSED" : "FAILED: credential send!");
            if (window.layoutTestController)
                layoutTestController.notifyDone();
        }
    };
    xhr.send();
} catch(e) {
    log("FAILED: got exception " + e.message);
}
</script>
</body>
</html>
