LayoutTests/http/tests/xmlhttprequest/access-control-preflight-credential-async-expected.txt

CONSOLE MESSAGE: https://localhost:8443/cookies/resources/cookie-utility.py?queryfunction=setFooCookie
CONSOLE MESSAGE: XHR response - Set the foo cookie
Test case for bug 37781: [XHR] Cross-Origin asynchronous request with credential raises NETWORK_ERR

PASSED
